pop-server/pyproject.toml
[tool.poetry]
name = "pop-server"
version = "0.4.0"
description = "The Precision Oncology Program (POP) is a joint at the University Hospital Zurich (USZ) to capture clinical data from oncology patients in a structured manner. The final aim of the project is to enable aggregated data analysis interactively."
readme = "README.md"
license = "MIT"
authors = ["Luis Fabregas-Ibanez <Luis.Fabregas-Ibanez@usz.ch>"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python"
]
packages = [{include = "pop-server"}]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
# Core dependencies
python = "~=3.11"
django = "5.1"
psycopg = "3.2.3"
psycopg2 = "2.9.10"
# Django extensions/apps
django-environ = "0.9.0"
django_extensions = "3.2.1"
django-secured-fields = "0.4.2"
django-measurement = "3.2.4"
django-ninja-extra = "0.21.6"
django-ninja-jwt = "5.3.5"
django-cors-headers = "4.6.0"
# Tools
requests = "2.32.0"
fhir-resources = "7.1.0" 
cachetools = "5.5.0"
tqdm = "4.67.0"
# Testing
pytest-django = "4.5.2"
factory_boy = "3.3.1"
parameterized = "0.9.0"
coverage = "7.2.3"
# WSGI server for production
gunicorn = "21.2.0"
# Serve dev-instance via HTTPS
Werkzeug = "3.0.0"
pyOpenSSL = "23.2.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "pop.settings"
python_files = [
    "pop/tests/**/test_*.py",
]
testpaths = ['pop/tests']
addopts = "--no-migrations"
django_find_project = "True"
django_debug_mode = "True"

[tool.coverage.run]
source = ["."]
omit = [
  "./.venv/*",
  "./venv/*",
  "*test*",
  "test_*",
  "*apps.py",
  "*manage.py",
  "*conftest.py",
  "*__init__.py",
  "*migrations*",
  "*asgi*",
  "*wsgi*",
  "*admin.py",
  "*urls.py",
  "**/migrations/*"
]

[tool.coverage.report]
omit = [
  "./.venv/*",
  "./venv/*",
  "*test*",
  "*apps.py",
  "*manage.py",
  "*conftest.py",
  "*__init__.py",
  "*migrations*",
  "*asgi*",
  "*wsgi*",
  "*admin.py",
  "*urls.py",
  "**/migrations/*"
]

